Move customcode implementation into extension
